Double to decimal change in Market.API

+++ Market.API/Application/DTOs/BuyCoinDto.cs
@@ -4,6 +4,6 @@ public class BuyCoinDto
 {
     public Guid UserId { get; set; }
     public string Symbol { get; set; } = string.Empty;
-    public double BuyPrice { get; set; }
-    public double BuyAmount { get; set; }
+    public decimal BuyPrice { get; set; }
+    public decimal BuyAmount { get; set; }
 }
