+++ Market.API/Application/Interfaces/ICoinService.cs
@@ -4,7 +4,7 @@ namespace Market.API.Application.Interfaces;
 
 public interface ICoinService
 {
-    Task<string> AddCoin(string Name, string Symbol, double Price, double MarketCap);
+    Task<string> AddCoin(string Name, string Symbol, decimal Price, decimal MarketCap);
     Task<List<CoinDto>> GetAllCoins();
     Task<CoinDto> GetCoinBySymbol(string Symbol);
     void BuyCoin(BuyCoinDto buyCoinDto);
