+++ Market.API/Domain/Entities/Coin.cs
@@ -7,8 +7,8 @@ public class Coin : BaseEntity
 {
     public string Name { get; set; } = string.Empty;
     public string Symbol { get; set; } = string.Empty;
-    public double CurrentPrice { get; set; }
-    public double MarketCap { get; set; }
+    public decimal CurrentPrice { get; set; }
+    public decimal MarketCap { get; set; }
 
     [BsonDateTimeOptions(Kind = DateTimeKind.Utc)]
     public DateTime LastUpdated { get; set; } = DateTime.UtcNow;
