+++ Market.API/Models/AddCoinRequest.cs
@@ -4,7 +4,7 @@ public class AddCoinRequest
 {
     public string Name { get; set; } = string.Empty;
     public string Symbol { get; set; } = string.Empty;
-    public double Price { get; set; } 
-    public double MarketCap { get; set; }
+    public decimal Price { get; set; } 
+    public decimal MarketCap { get; set; }
 
 }
